requirements.txt
google-api-python-client==2.108.0
google-auth==2.23.4
pydantic-settings==2.1.0